pyproject.toml
[tool.poetry]
name = "ran.routing.tti"
version = "0.1.0"
description = ""
authors = ["Everynet"]

[tool.poetry.dependencies]
python = "^3.10"
protobuf = "<3.21"
# We don't care about this package versions, because they are hardly depend on "protobuf" version
grpcio = "*" 
grpcio-tools = "*" 

structlog = "^22.3.0"
paho-mqtt = "^1.6.1"
asyncio-mqtt = "^0.16.1"
aiohttp = "^3.8.3"
environs = "^9.5.0"
pydantic = {extras = ["dotenv"], version = "^1.10.2"}

pylorawan = {git = "https://github.com/everynet/pylorawan.git", rev = "main"}
ran = {git = "https://github.com/everynet/ran.routing.pyclient.git", rev = "main"}

# TODO: proper installation for tti_contracts
tti_contracts = {path = "./contracts", develop = true}


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
pytest-asyncio = "^0.20.3"
black = "^22.12.0"
isort = "^5.11.2"
bandit = "^1.7.4"
flake8 = "^6.0.0"
flake8-bugbear = "^22.12.6"
flake8-isort = "^5.0.3"
flake8-bandit = "^4.1.1"
mypy = "^0.991"


[tool.bandit]
targets = ["src", "tests"]

[tool.isort]
profile = "black"
line_length = 120

[tool.black]
line-length = 120
target-version = ["py39"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.pip-cache
  | \.poetry
  | \.venv
  | \.vscode
  | _build
  | buck-out
  | build
  | dist
  | __pycache__
)/
'''

[build-system]
requires = ["poetry-core>=1.3.0"]
build-backend = "poetry.core.masonry.api"
